imgen/requirements.txt
diffusers
transformers
accelerate
torch
opencv-python
pillow
protobuf
scipy
controlnet_aux
peft
